+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - feat: add `box3.expandByMargin` to expand a Box3 by a uniform margin in all directions
 - feat: rename `ray3` / `Ray3` to `raycast3` / `Raycast3` to better reflect its purpose as a ray used for raycasting, including a length component
 - feat: add `box3.surfaceArea` to compute the surface area of a Box3
+- feat: add `box3.empty` to set a Box3 to an empty state (min at +Infinity, max at -Infinity)
 
 ## 0.0.6